solution/feedster_server/requirements.txt
grpcio==1.29.0
grpcio-tools==1.29.0
protobuf==3.12.1
six==1.14.0
git+https://github.com/markeyev/feedparser_wrapper.git
